+++ Lab2/objects/TestObject.cs
@@ -9,12 +9,12 @@ namespace Lab2.objects
     public class TestObject
     {
         public decimal Iter { get; set; }
-        public decimal N { get; set; }
-        public decimal pk { get; set; }
-        public decimal pm { get; set; }
-        public decimal T { get; set; }
-        public decimal AvgMark { get; set; }
-        public decimal MinMark { get; set; }
-        public decimal MaxMark { get; set; }
+        public double N { get; set; }
+        public double pk { get; set; }
+        public double pm { get; set; }
+        public double T { get; set; }
+        public double AvgMark { get; set; }
+        public double MinMark { get; set; }
+        public double MaxMark { get; set; }
     }
 }
